Move garbage collection into dag_gc.c
