+++ makeflow/src/Makefile
@@ -6,7 +6,7 @@ include ../../rules.mk
 LOCAL_LINKAGE=$(CCTOOLS_GLOBUS_LDFLAGS)
 
 EXTERNAL_DEPENDENCIES = ../../work_queue/src/libwork_queue.a ../../chirp/src/libchirp.a ../../dttools/src/libdttools.a
-OBJECTS = dag.o dag_node.o dag_file.o dag_variable.o dag_task_category.o dag_visitors.o makeflow_common.o lexer.o parser.o
+OBJECTS = dag.o dag_node.o dag_file.o dag_variable.o dag_task_category.o dag_visitors.o dag_log.o dag_gc.o makeflow_common.o lexer.o parser.o
 PROGRAMS = makeflow makeflow_viz makeflow_analyze makeflow_linker
 SCRIPTS = condor_submit_makeflow makeflow_graph_log makeflow_monitor starch makeflow_linker_perl_driver makeflow_linker_python_driver
 TARGETS = $(PROGRAMS)
@@ -14,6 +14,8 @@ TARGETS = $(PROGRAMS)
 all: $(TARGETS)
 
 makeflow makeflow_viz makeflow_analyze: $(OBJECTS)
+
+makeflow: makeflow_summary.o
 
 $(PROGRAMS): $(EXTERNAL_DEPENDENCIES)
